server/pyproject.toml

[tool.poetry]
name = "ToDo List API"
version = "0.0.1"
description = "A simple Django-based to-do list API"
authors = ["Ruslan Iliev <dddrsln@mail.ru>"]

[tool.poetry.dependencies]
python = ">=3.12"
django = "^5.2.1"
black = "^25.1.0"
psycopg2-binary = "^2.9.10"
django-extensions = "^4.1"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
